refactor: Extract mergeWithCustomize and drop mergeMultiple
It's better to separate mergeWithCustomize as that's better for typing.
Now normal merge uses this below.

I ended up dropping mergeMultiple as it feels it's a special case that's
better handled in user space.
