--- test/merge-multiple.test.ts
@@ -1,6 +1,0 @@
-import { mergeMultiple } from "../src";
-import mergeMultipleTests from "../helpers/merge-multiple-tests";
-
-describe("Multiple merge", function () {
-  mergeMultipleTests(mergeMultiple);
-});
